feat(startproject): add support to indicate Godot project directory.
Add --godot-dir option on startproject command to specify the Godot
project directory where the .gdextension file will be saved. It's useful
in case you want to separate your godot project from the rust project.
It also specifies the libraries paths as well so this doesn't need to be
tweaked by the user.